+++ CHANGELOG.md
@@ -7,9 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0]
+
+### Changed
+
+- Certificate is now retrieved from the Keyvault
+- Robot is now pointed at real Digital Post service
+
 ## [1.0.0]
 
 - Initial release
 
-[Unreleased]: https://github.com/itk-dev-rpa/udtraek-tilmelding-digital-post/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/itk-dev-rpa/udtraek-tilmelding-digital-post/compare/1.1.0...HEAD
+[1.1.0]: https://github.com/itk-dev-rpa/udtraek-tilmelding-digital-post/compare/1.1.0...HEAD
 [1.0.0]: https://github.com/itk-dev-rpa/udtraek-tilmelding-digital-post/compare/1.0.0...HEAD